correct table models to simplify work with data base
